requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai==0.3.2
supabase==2.3.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
pyjwt==2.8.0
cryptography==42.0.0
